Moved common camera test resources to external files
