+++ unit_tests/base/mock/mock.cpp
@@ -6,6 +6,7 @@
 #include "mock/HasStateMock.hpp"
 #include "mock/ImtqDriverMock.hpp"
 #include "mock/ImtqTelemetryCollectorMock.hpp"
+#include "mock/LineIOMock.hpp"
 #include "mock/experiment.hpp"
 #include "mock/fm25w.hpp"
 #include "mock/n25q.hpp"
@@ -102,5 +103,13 @@ N25QDriverMock::N25QDriverMock()
 {
 }
 N25QDriverMock::~N25QDriverMock()
+{
+}
+
+LineIOMock::LineIOMock()
+{
+}
+
+LineIOMock::~LineIOMock()
 {
 }
